requirements.txt
asgiref==3.8.1
attrs==24.2.0
autopep8==2.3.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1
django-cors-headers==4.4.0
django-currentuser==0.8.0
django-environ==0.11.2
django-phonenumber-field==8.0.0
django-pytest==0.2.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
environ==1.0
flake8==7.1.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.1
phonenumbers==8.13.44
phonenumberslite==8.13.44
pillow==10.4.0
pluggy==1.5.0
psycopg==3.2.3
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
django-filter==24.3
uritemplate==4.1.1
urllib3==2.2.3
elasticsearch==8.15.1
django-elasticsearch-dsl==8.0